display multiple images in TableView
